Move type generation to a worker thread
