+++ packages/overlay/scripts/utils.ts
@@ -1,0 +1,5 @@
+import path from 'path'
+import { fileURLToPath } from 'url'
+
+export const __filename = fileURLToPath(import.meta.url)
+export const __dirname = path.dirname(__filename)
